+++ src/variables/aboutConsts.js
@@ -3,6 +3,19 @@ const AUTHOR = "Arnav Thorat";
 
 const VERSION = "0.6.9";
 const BUILD = "69";
-const IN_BETA = true;
 
-module.exports = { GLOBAL_NAME, AUTHOR, VERSION, BUILD, IN_BETA };
+const IN_BETA = true;
+const TIMEBOMB_ACTIVATED = true;
+const EXPIRY_DATE = new Date(2023, 4, 19);
+const TIMEBOMB_COUNT = 90;
+
+module.exports = {
+  GLOBAL_NAME,
+  AUTHOR,
+  VERSION,
+  BUILD,
+  IN_BETA,
+  TIMEBOMB_ACTIVATED,
+  EXPIRY_DATE,
+  TIMEBOMB_COUNT,
+};