+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This is the official BubbleOS `CHANGELOG`! All features will be recorded in _'gr
 
 - Added a function in the `symlink` command which creates a symbolic link! (You can use `-c` to check if it is a symbolic link.)
 - Added the ability to use the `-y` flag on the `del` command to skip the confirmation prompt.
+- Added a timebomb to BubbleOS.
+- The `ls` command will now recognize symbolic files/folders and color them accordingly (red).
 
 ### Changed/Fixed Features